Fixes #171.
When the page was reloaded, the dashboard controller tried to access
the user profile before it was returned from the API. This resulted into
a JavaScript error and prevented the rest of the dashboard functionality.

This change adds an event to be broadcasted from homeCtrl.js that is
listened for in case the user profile is not yet fetched.