Fix the UI flow after a fresh login.
When entering the home page after a login, the ADAL.js library hasn't
yet fully processed the login result when the code in the home controller
is executed. That is why we need to also subscribe to the login event
so that we can react once the login is processed.