Switch endpoint from which groups are fetched.
In the group manager view, the groups used to be fetched using the same API
that the dashboard uses. That works, but returns some amount of data that is
not needed in the group manager.

This commit shouldn't change what the end user sees, but aligns better the API
endpoints that the group management functionality uses.