Avoid multiple concurrent profile fetches.
In certain page reload and route change combinations, the profile fetch
was initiated twice very close to each other. In the backend, this created
a race condition where the first fetch triggered a new user to start being
created while the second fetch came in before that had completed. The end
result was a key duplication error in the documentdb, because the second
fetch tried to start creating exactly the same kind of document.

This doesn't fix the root cause and the race condition is still possible,
but after this commit, the end user shouldn't see an issue in normal
usage.

Fixes #257.